detection: shrink and split footprints
This throws away masked pixels around the edge of footprints, and
then breaks up any non-contiguous remnants into separate footprints.
This deals with footprints from two different objects on either side
of a chip or exposure gap (i.e., on a warp or coadd) where the
interpolation has left a ridge of high (but masked) pixels between
them.

We might prefer to change the FootprintSet constructor to respect
masks, but we want to keep the grow because that prevents us from
splitting objects that are saturated in the center.