PixelFlags: make additional flags configurable
We recently added the CLIPPED mask plane to the pixel flags, but
it was only going into the schema if it was defined. This produced
different schemas, depending on whether the mask plane was defined,
which can lead to subtle problems.

Instead of hard-wiring a flag based on a mask plane that may or may
not exist, this change makes the list of mask planes (besides the
defaults) configurable. That means that the schema will always be
the same (when built with the same configuration). This also allows
for easy extension of the pixel flags.

The hard-coded flags could now (mostly) be removed, except that would
rename them, and they are referred to in other parts of the code that
we don't want to track down right now.