CurveOfGrowth: split out result of calculation into a separate class
The CurveOfGrowth object can't be pickled: it holds things like
Keys to a SourceCatalog and individual SourceRecords which aren't
picklable. However, we desire to pickle the results of the curve
of growth calculation so we can store it or transfer it.  So we
split off the result into its own (picklable) class that can be
used separately. Methods in CurveOfGrowth forward to
CurveOfGrowthResults, and the results object is built on demand
and cached (with @property).

Added a test to ensure the results can be pickled.