Improve calculation of photon noise in aperture corrections.
The errors on the aperture correction should be purely systematic: we want
to subtract off any expected noise from photon shot noise.  This change
improves our calculated of that expected noise by correctly taking into
account the covariances between flux measurements even when we're not
dealing with aperture fluxes and/or background-limited data.